feat: move logic to atom setter
