+++ apps/namadillo/src/App/Settings/Advanced.tsx
@@ -1,7 +1,6 @@
 import { ActionButton, Input, Stack } from "@namada/components";
 import SettingsRoute from "App/Settings/routes";
 import { indexerUrlAtom, rpcUrlAtom } from "atoms/settings";
-import { useValidateApiUrl } from "hooks/useValidateApiUrl";
 import { useAtom } from "jotai";
 import { useState } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
@@ -9,7 +8,6 @@ import { useLocation, useNavigate } from "react-router-dom";
 export const Advanced = (): JSX.Element => {
   const navigate = useNavigate();
   const location = useLocation();
-  const validateApiUrl = useValidateApiUrl();
   const [currentRpc, setCurrentRpc] = useAtom(rpcUrlAtom);
   const [currentIndexer, setCurrentIndexer] = useAtom(indexerUrlAtom);
   const [rpc, setRpc] = useState(currentRpc);
@@ -22,8 +20,8 @@ export const Advanced = (): JSX.Element => {
     e.preventDefault();
     setValidatingUrl(true);
     const [rpcResult, indexerResult] = await Promise.allSettled([
-      validateApiUrl(rpc),
-      validateApiUrl(indexer),
+      setCurrentRpc(rpc),
+      setCurrentIndexer(indexer),
     ]);
     if (rpcResult.status === "rejected") {
       setRpcError(String(rpcResult.reason));
@@ -35,8 +33,6 @@ export const Advanced = (): JSX.Element => {
       rpcResult.status === "fulfilled" &&
       indexerResult.status == "fulfilled"
     ) {
-      setCurrentRpc(rpcResult.value);
-      setCurrentIndexer(indexerResult.value);
       navigate(SettingsRoute.index(), { replace: true, state: location.state });
     }
     setValidatingUrl(false);
